FFWEB-2405: add unit tests for forward slash removing methods
